+++ src/actinia_core/rest/user_api_key.py
@@ -35,7 +35,7 @@ from flask_restful_swagger_2 import swagger
 from flask import jsonify, make_response, g
 from actinia_api.swagger2.actinia_core.schemas.user_api_key import TokenResponseModel
 
-from .base_login import LoginBase
+from actinia_core.rest.base.base_login import LoginBase
 from actinia_core.core.common.app import auth
 from actinia_core.core.common.api_logger import log_api_call
